Fix a problem with try-with-resources and JavaNCSS

+++ src/main/java/org/assertj/db/util/Values.java
@@ -352,7 +352,7 @@ public class Values {
         } else if (number instanceof Integer) {
           actual = ((Integer) number).longValue();
         } else if (number instanceof Long) {
-          actual = ((Long) number).longValue();
+          actual = (Long) number;
         }
 
         if (actual != null && actual.longValue() == Long.parseLong(expected)) {
